+++ persistent/ChangeLog.md
@@ -1,14 +1,45 @@
 # Changelog for persistent
 
-## unreleased
+## 2.14.6.0 (unreleased)
 
-* [#1460] https://github.com/yesodweb/persistent/pull/1468
+* [#1503](https://github.com/yesodweb/persistent/pull/1503)
+    * Create Haddocks from entity documentation comments
+* [1497](https://github.com/yesodweb/persistent/pull/1497)
+    * Always generates `SymbolToField "id"` instance
+* [#1509](https://github.com/yesodweb/persistent/pull/1509)
+    * Provide `ViaPersistEntity` for defining `PathMultiPiece` for entity keys.
+
+## 2.14.5.2
+
+* [#1513](https://github.com/yesodweb/persistent/pull/1513)
+    * Support GHC 9.8 and `aeson-2.2`
+
+## 2.14.5.1
+
+* [#1496](https://github.com/yesodweb/persistent/pull/1496)
+    * Fixes name shadowing error at the generated `keyFromRecordM` function.
+* [#1505](https://github.com/yesodweb/persistent/pull/1505)
+    * Fixes the comment line parsing rule so that accommodates paragraph breaks.
+
+## 2.14.5.0
+
+* [#1469](https://github.com/yesodweb/persistent/pull/1469)
+    * Change default implementation for `insertUnique_` to not perform
+      unecessary queries (mirrors 1449)
+* [#1437](https://github.com/yesodweb/persistent/pull/1437)
+    * Add `existsBy` to `PersistUniqueRead`
+
+## 2.14.4.5
+
+* [#1460](https://github.com/yesodweb/persistent/pull/1468)
     * Remove extraneous `map toPersistValue` call in the `mkInsertValues`
       function, as it evaluates to `id`.
+* [#1476](https://github.com/yesodweb/persistent/pull/1476)
+    * Fix `mkRecordName` to suffix `_` if the field name matches any of Haskell keywords.
 
 ## 2.14.4.4
 
-* [#1460] https://github.com/yesodweb/persistent/pull/1460
+* [#1460](https://github.com/yesodweb/persistent/pull/1460)
     * Fix a problem where a `Primary` key causes `mkPersist` to generate code
       that doesn't compile under `NoFieldSelectors`